#93: [Refactor] navigationのロジックをカスタムフックに分離
