+++ frontend-nodejs/app/hooks/useSettlementNavigation.ts
@@ -1,0 +1,58 @@
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+
+/**
+ * セトルメントページのナビゲーション管理のためのカスタムフック
+ * ページ遷移とツールチップ表示を管理
+ * @returns ナビゲーションに関する状態とハンドラ
+ */
+export function useSettlementNavigation() {
+  const router = useRouter();
+  const [showStoreTooltip, setShowStoreTooltip] = useState(false);
+  const [showGiftTooltip, setShowGiftTooltip] = useState(false);
+
+  /**
+   * ルーム作成ページへ移動
+   */
+  const goToCreateRoom = () => {
+    router.push("/host/building-selection");
+  };
+
+  /**
+   * ルーム参加ページへ移動
+   */
+  const goToJoinRoom = () => {
+    router.push("/join");
+  };
+
+  /**
+   * ストアページへ移動（現在は開発中メッセージを表示）
+   */
+  const goToStore = () => {
+    // 将来的にはストアページへのルーティングを実装
+    alert(
+      "ストアは開発中です！今後さまざまな建物やアイテムを購入できるようになります。",
+    );
+  };
+
+  /**
+   * プレゼントページへ移動（現在は開発中メッセージを表示）
+   */
+  const goToGifts = () => {
+    // 将来的にはプレゼントページへのルーティングを実装
+    alert(
+      "プレゼントボックスは開発中です！今後様々な報酬を受け取れるようになります。",
+    );
+  };
+
+  return {
+    showStoreTooltip,
+    showGiftTooltip,
+    setShowStoreTooltip,
+    setShowGiftTooltip,
+    goToCreateRoom,
+    goToJoinRoom,
+    goToStore,
+    goToGifts,
+  };
+}
